Added more plugins
fzf for fuzzy file finding, splitjoin for multiline expansion and
contractions, vim-airline for nice status bar, tagbar for nice tags
associated with classes and functions
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
    modified:   .gitmodules
    new file:   bundle/fzf
    new file:   bundle/fzf.vim
    new file:   bundle/splitjoin.vim
    new file:   bundle/tagbar
    new file:   bundle/vim-airline
    modified:   vimrc

Changes not staged for commit:
    modified:   bundle/jupyter-vim (untracked content)
    modified:   bundle/vim-calc (untracked content)